improve contracts
