+++ contract/README.md
@@ -1,7 +1,7 @@
 # Hardhat Project
 These are the smart contracts for a lending and borrowing platform written in Solidity and deployed onto the Linea network. 
 Users will be able to borrow tokens, using an NFT as a collateral.
-Smart contract testing was done using Remix.
+Smart contract tested using Remix.
 
 ```shell
 npx hardhat help
@@ -10,7 +10,6 @@ REPORT_GAS=true npx hardhat test
 npx hardhat node
 npx hardhat ignition deploy ./ignition/modules/SimpleLending.js --network linea_sepolia
 ```
-
 
 # Deployments
 Linea Sepolia testnet: 
@@ -22,5 +21,5 @@ Pup NFT contract address: [0xc1EF9b45Cd110ad11bae39A173BDaE161E240557](https://s
 - Add access control to only allow admins to withdraw the NFT collateral.
 - Incorporate more tokens and liquidity pools.
 - Allow P2P lending and borrowing.
-- Using an Oracle to determine the value of an NFT.
+- Use an Oracle to determine the value of an NFT.
 - Allow bridging of ETH and ERC-20 tokens from Ethereum to Linea network.
