+++ contract/contracts/MockToken.sol
@@ -5,11 +5,11 @@ pragma solidity ^0.8.20;
 import "@openzeppelin/contracts/token/ERC20/ERC20.sol";
 import "@openzeppelin/contracts/access/Ownable.sol";
 
-contract SimpleToken is ERC20, Ownable {
+contract MockToken is ERC20, Ownable {
     constructor(
         address initialOwner
-    ) ERC20("Bone", "BONE") Ownable(initialOwner) {
-        _mint(msg.sender, 1000000000 * 10 ** decimals());
+    ) ERC20("MockUSDC", "USDC") Ownable(initialOwner) {
+        _mint(msg.sender, 1000000000 * 10 ** decimals()); // 1 billion tokens
     }
 
     function mint(address to, uint256 amount) public onlyOwner {
